// src/app/core/services/model/parameterDefinitionDto.ts
/**
 * TWMS REST API
 * TWMS REST API
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AccessAuthorizationDto } from './accessAuthorizationDto';
import { ReadWriteDto } from './readWriteDto';
import { TwmsFeaturesDto } from './twmsFeaturesDto';

export interface ParameterDefinitionDto { 
    deviceType?: string;
    keyName?: string;
    name?: string;
    abbreviation?: string;
    remark?: string;
    twmsFeatures?: TwmsFeaturesDto;
    viptool?: string;
    readWrite?: ReadWriteDto;
    accessAuthorization?: AccessAuthorizationDto;
    object?: string;
    dataType?: string;
    enumOptions?: Array<string>;
    unit?: string;
    min?: number;
    max?: number;
    defaultValue?: string;
    selectableOptions?: Array<string>;
}