// src/app/core/services/model/deviceStatusSimple.ts
/**
 * TWMS REST API
 * TWMS REST API
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export type DeviceStatusSimple = 'Online' | 'Offline';

export const DeviceStatusSimple = {
    Online: 'Online' as DeviceStatusSimple,
    Offline: 'Offline' as DeviceStatusSimple
};